# [CSL-2132] Optimize slog apply by batching