--- a/lib/src/Pos/Launcher/Mode.hs
+++ b/lib/src/Pos/Launcher/Mode.hs
@@ -31,7 +31,7 @@ import           Mockable.Production (Production)
 import           Pos.Core (HasConfiguration, Timestamp)
 import           Pos.DB (NodeDBs)
 import           Pos.DB.Block (dbGetSerBlockRealDefault, dbGetSerUndoRealDefault,
-                               dbPutSerBlundRealDefault)
+                               dbPutSerBlundsRealDefault)
 import           Pos.DB.Class (MonadDB (..), MonadDBRead (..))
 import           Pos.DB.Rocks (dbDeleteDefault, dbGetDefault, dbIterSourceDefault, dbPutDefault,
                                dbWriteBatchDefault)
@@ -86,7 +86,7 @@ instance HasConfiguration => MonadDB InitMode where
     dbPut = dbPutDefault
     dbWriteBatch = dbWriteBatchDefault
     dbDelete = dbDeleteDefault
-    dbPutSerBlund = dbPutSerBlundRealDefault
+    dbPutSerBlunds = dbPutSerBlundsRealDefault
 
 instance (HasConfiguration, HasInfraConfiguration, MonadSlotsData ctx InitMode) =>
          MonadSlots ctx InitMode