Copying from YT:

I've managed to speed up application by 30%. I was connecting local node to the mainnet directly, built with -O2 --no-asserts, of course. For the default 2.1k blocks batch it was about 530ms, now it's 390ms. I don't think I can optimize further here, the main bottleneck seems to be the fact we store blocks in files instead of rocksdb. I must admit though, that for now the biggest problem seems to be the server side. Average main retrieval loop length is 36sec, of which verification (850ms) and application (450ms) plus extra related local thing takes about 1.4s. other 34.5 sec are diffusion.getblocks. My networking is good, so i suppose the thing is that server nodes are terribly slow.